Better separation of editors for different languages, which causes fewer code clones.
