+++ tools/Graph Explorer/GraphExplorer/SocratexGraphExplorer.Vis.netcore.csproj
@@ -372,6 +372,7 @@
       <ProjectReference Include="..\Plugins\XppProgrammingPlugins.netcore.csproj" />
       <ProjectReference Include="..\SocratexGraphExplorer.Common.netcore\SocratexGraphExplorer.Common.netcore.csproj" />
       <ProjectReference Include="..\SocratexGraphExplorer.DefaultRenderers.netcore\SocratexGraphExplorer.DefaultRenderers.netcore.csproj" />
+      <ProjectReference Include="..\SocratexGraphExplorer.SourceEditor.netcore\SocratexGraphExplorer.SourceEditor.netcore.csproj" />
     </ItemGroup>
     <ItemGroup>
       <Compile Update="Properties\Settings.Designer.cs">
